feat: add cli argument to render a cluster in non interactive mode
